requirements.txt
mcp
tushare
pandas
python-dotenv

